requirements.txt
pandas
matplotlib
numpy
scikit-learn
python-dotenv
mysql-connector-python
pymysql
